Introduce GamePersistenceService
