+++ src/main/kotlin/de/smartsquare/kickchain/kickway/storing/GameController.kt
@@ -1,9 +1,6 @@
 package de.smartsquare.kickchain.kickway.storing
 
 import de.smartsquare.kickchain.kickway.Blockchain
-import de.smartsquare.kickchain.kickway.elo.EloRating
-import de.smartsquare.kickchain.kickway.elo.EloRatingRepository
-import de.smartsquare.kickchain.kickway.elo.readjust
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
@@ -11,8 +8,7 @@ import org.springframework.web.bind.annotation.RestController
 
 @RestController
 class GameController(
-    private val gameRepository: GameRepository,
-    private val eloRatingRepository: EloRatingRepository
+    private val gamePersistenceService: GamePersistenceService
 ) {
 
     @PostMapping("/game")
@@ -30,15 +26,7 @@ class GameController(
             return ResponseEntity.badRequest().body("A team cannot score minus goals")
         }
 
-        val eloOne = eloRatingRepository.findEloRatingByTeamFirstAndTeamSecond(game.team1.first, game.team1.second)
-            .orElse(EloRating(game.team1.first, game.team1.second, elo = 1000.0, matches = 0))
-        val eloTwo = eloRatingRepository.findEloRatingByTeamFirstAndTeamSecond(game.team2.first, game.team2.second)
-            .orElse(EloRating(game.team2.first, game.team2.second, elo = 1000.0, matches = 0))
-        val (newEloOne, newEloTwo) = game.readjust(eloOne, eloTwo)
-        eloRatingRepository.save(newEloOne)
-        eloRatingRepository.save(newEloTwo)
-
-        gameRepository.save(game)
+        gamePersistenceService.store(game)
 
         return ResponseEntity.ok().build()
     }
